% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barplot.R
\name{barplot_gsea}
\alias{barplot_gsea}
\title{Barplot of scripts using fgsea algorithm}
\usage{
barplot_gsea(
  height,
  x = "NES",
  color = "p.adjust",
  showCategory = 8,
  font.size = 12,
  title = "",
  xlab = NULL,
  ylab = NULL,
  orderBy = "x",
  decreasing = TRUE,
  ...
)
}
\arguments{
\item{countdata}{Raw counts table with entrez or ensembl IDs (output from merged bam files)}
}
\value{
Barplot from GSEA scrip
}
\description{
Barplot of scripts using fgsea algorithm
}
