% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversions_human.R
\name{convertHumanGeneList}
\alias{convertHumanGeneList}
\title{Easily convert human to mouse genes}
\usage{
convertHumanGeneList(x)
}
\arguments{
\item{x}{input list of human genes}
}
\value{
list of input genes with respective conversions
}
\description{
Easily convert human to mouse genes
}
