% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_conversions.R
\name{convertMouseGeneList}
\alias{convertMouseGeneList}
\title{Easily convert between mouse, human and zebrafish genes}
\usage{
convertMouseGeneList(x)
}
\arguments{
\item{x}{input list of genes}
}
\value{
list of input genes with respective conversions
}
\description{
Easily convert between mouse, human and zebrafish genes
}
