% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Corrplot.R
\name{corrplot}
\alias{corrplot}
\title{Interactive correlation plot}
\usage{
corrplot(frame1, frame2, com_col, FCcol, FDRcol, namex = NULL, namey = NULL)
}
\arguments{
\item{frame1}{1st data frame}

\item{frame2}{2nd data frame}

\item{com_col}{Common column to match rows (ie. gene symbols)}

\item{FCcol}{label common fold change columns}

\item{FDRcol}{label common FDR/adj P value columns}

\item{namex}{name for x axis}

\item{namey}{name for y axis}
}
\value{
plotly interactive correlation plot of input dataframes
}
\description{
using two data tables, generate interactive correlation plot
}
