% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPMtable.R
\name{CPMtable}
\alias{CPMtable}
\title{Return table of filtered and TMM (edgeR) normalised CPMs}
\usage{
CPMtable(
  countdata,
  sampleinfo,
  thresh = 0.5,
  keep,
  group_param,
  species,
  genekeytype,
  archived = "FALSE",
  log = F
)
}
\arguments{
\item{countdata}{Raw counts table with entrez or ensembl IDs (output from merged bam files)}

\item{sampleinfo}{Sampleinfo file (see template)}

\item{thresh}{Minimum CPM threshold to exclude}

\item{keep}{Minimum number of samples below thresh before exclusion (value usually set to number of replciates per group)}

\item{group_param}{select group_param column to compare from sampleinfo}

\item{species}{"human" or "mouse"}

\item{genekeytype}{"entrez" or "ensembl"}

\item{archived}{Should be set to FALSE, unless using archived Human Ensembl database}

\item{log}{Set to true to output normalised LogCPM}
}
\value{
Table of normalised and filtered CPMs with converted gene symbols
}
\description{
This function takes raw countdata+sampleinfo+group_pram and filters on minimum CPM and sample replicates. The genekeytype, "ensembl" or "entrez" and species, "human" or "mouse" must be specified.
}
