% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot.R
\name{dotplot_gsea}
\alias{dotplot_gsea}
\title{DOTplot for GSEA scripts using fgsea algorithm}
\usage{
dotplot_gsea(
  object,
  x = "geneRatio",
  color = "p.adjust",
  showCategory = 20,
  size = NULL,
  split = NULL,
  font.size = 12,
  title = "",
  orderBy = "x",
  decreasing = TRUE
)
}
\arguments{
\item{countdata}{Raw counts table with entrez or ensembl IDs (output from merged bam files)}
}
\value{
Dotplot of results from GSEAscript
}
\description{
DOTplot for GSEA scripts using fgsea algorithm
}
