% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSEAscripts.R
\name{gseatest}
\alias{gseatest}
\title{GSEA scripts using fgsea algorithm}
\usage{
gseatest(
  input,
  method,
  species,
  orderby = "logFC",
  cat = NULL,
  subcat = NULL,
  nPerm = 1000,
  minGSSize = 15,
  maxGSSize = 500,
  pvalueCutoff = 0.05,
  by = "fgsea",
  plotpathway = NULL,
  expname,
  DEgroup,
  showCategory = 20,
  customGS = NULL,
  plot = F,
  ...
)
}
\arguments{
\item{countdata}{Raw counts table with entrez or ensembl IDs (output from merged bam files)}
}
\value{
Table of normalised and filtered DGElist objects for limma/voom and EdgeR
}
\description{
GSEA scripts using fgsea algorithm
}
