% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap.R
\name{heatmapscript}
\alias{heatmapscript}
\title{Heatmapscript use heatmapscriptnew if error}
\usage{
heatmapscript(
  data = NULL,
  selcol = NULL,
  refno = 1,
  refgene = data,
  name = NULL,
  destination = ..,
  clcol = TRUE
)
}
\arguments{
\item{data}{CPM table with entrez or ensembl IDs (output of sekrnaseq::CPMlist function)}
}
\value{
Heatmaps outputed to heatmap/exp_name folder
}
\description{
Heatmapscript use heatmapscriptnew if error
}
