% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap2.R
\name{heatmapscriptnew}
\alias{heatmapscriptnew}
\title{Heatmapscriptnew use heatmapscript if error}
\usage{
heatmapscriptnew(
  data = NULL,
  selcol = NULL,
  refno = 1,
  refgene = data,
  name = NULL,
  destination,
  clcol = TRUE
)
}
\arguments{
\item{data}{CPM table with entrez or ensembl IDs (output of sekrnaseq::CPMlist function)}
}
\value{
Heatmaps outputed to heatmap/exp_name folder
}
\description{
Heatmapscriptnew use heatmapscript if error
}
