% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_load.R
\name{install_load}
\alias{install_load}
\title{Load packages and install them if necessary from BiocManager}
\usage{
install_load(package1, ...)
}
\arguments{
\item{package1}{Provide list of packages to load or install}
}
\value{
List of packages that are loaded or installed
}
\description{
This function loads the provided list of packages, and installs them if necessary from BiocManager
}
