% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadgenes.R
\name{loadgenes}
\alias{loadgenes}
\title{loadgenes from MsigDB genesets}
\usage{
loadgenes(path1, species)
}
\arguments{
\item{path1}{name of pathway}

\item{species}{a=mouse or b=human}
}
\value{
Table of normalised and filtered DGElist objects for limma/voom and EdgeR
}
\description{
loadgenes from MsigDB genesets
}
