% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAplot.R
\name{maplotniceqlf}
\alias{maplotniceqlf}
\title{Plotting RNA-seq data input from DE_analysis}
\usage{
maplotniceqlf(
  data_in,
  xaxis = "logCPM",
  yaxis = "logFC",
  name_output,
  exp_name
)
}
\arguments{
\item{data_in}{Output from DE_analysis}

\item{xaxis}{Input string for xaxis}

\item{yaxis}{Input string for yaxis}

\item{name_output}{Name to title plot}

\item{exp_name}{exp_name, also determine destination where file is saved}
}
\value{
ggplots of RNAseq data
}
\description{
Using output file from DE_analysis to generate volcano, MA, correlation and MDS plots.
}
