% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rastervolc.R
\name{rastervolc}
\alias{rastervolc}
\title{Plotting RNA-seq data input from DE_analysis}
\usage{
rastervolc(data_in, label_peaks, xaxis, yaxis, name_output, exp_name)
}
\arguments{
\item{data_in}{Output from DE_analysis}

\item{xaxis}{Input string for xaxis}

\item{yaxis}{Input string for yaxis}

\item{name_output}{Name to title plot}

\item{exp_name}{exp_name, also determine destination where file is saved}
}
\value{
ggplots of RNAseq data
}
\description{
Using output file from DE_analysis to generate volcano, MA, correlation and MDS plots.
}
