% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectgenes.R
\name{selectgenes}
\alias{selectgenes}
\title{selectgenes, load long list of all genesets from MsigDB genesets}
\usage{
selectgenes(species)
}
\arguments{
\item{species}{a=mouse or b=human}
}
\value{
Table of normalised and filtered DGElist objects for limma/voom and EdgeR
}
\description{
selectgenes, load long list of all genesets from MsigDB genesets
}
